refactor: LoadingSpinner
