+++ src/components/slider/slider.scss
@@ -17,7 +17,7 @@
   display: flex;
   justify-content: center; /* 슬라이드를 중앙 정렬 */
   align-items: center;
-  font-family: "pretendard", sans-serif;
+
   color: $color-white-1;
 }
 
@@ -33,7 +33,7 @@
   display: flex;
   justify-content: center; /* 슬라이드를 중앙 정렬 */
   align-items: center;
-  font-family: "pretendard", sans-serif;
+
   color: $color-white-1;
 }
