+++ src/styles/global.scss
@@ -1,6 +1,6 @@
 @import "./font";
 @import "./color";
-@import "../components/slider/tributeSlider.scss";
+@import "../components/slider/slider.scss";
 @import "../components/loadingSpinner/loadingSpinner.scss";
 
 *,
